% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajout.R
\name{ajout}
\alias{ajout}
\title{Ajout}
\usage{
ajout(pp = "dat")
}
\arguments{
\item{pp}{the pattern you want to recognize in the name of the files you want to work on.}
}
\description{
Ajout
}
\details{
this function is made to add ring limite by clicking on the profile where you want to position it.
}
