% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ajout_branches.R
\name{ajout_branch}
\alias{ajout_branch}
\title{Ajout_branches}
\usage{
ajout_branch(pp = "dat", milmax)
}
\arguments{
\item{pp}{the pattern you want to recognize in the name of the files you want to work on.}
}
\description{
Ajout_branches
}
