% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajustepositionlimite.R
\name{ajustepositionlimite}
\alias{ajustepositionlimite}
\title{Ajustepositionlimit}
\usage{
ajustepositionlimite(uu = "dat", intliss1, intliss2)
}
\arguments{
\item{uu}{the pattern you want to recognize in the name of the files you want to work on.}

\item{intliss1}{Intensity of the first smoothing function.}

\item{intliss1}{Intensity of the seconde smoothing function.}
}
\description{
Ajustepositionlimit
}
