% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul.mat.bt.r
\name{calcul.mat.bt}
\alias{calcul.mat.bt}
\title{Calcul.mat.bt}
\usage{
calcul.mat.bt(ww = "dat", papa, subst1, subst2)
}
\arguments{
\item{ww}{the pattern you want to recognize in the name of the files you want to used to calculate variables.}

\item{papa}{indicate the value of the interval of resolution in Windendro.}

\item{subst1}{where to start in the substring of the file name to determined the identification code}

\item{subst2}{where to stop in the substring  of the file name to determined the identification code}
}
\description{
Calcul.mat.bt
}
