% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_branche.R
\name{cut_branche}
\alias{cut_branche}
\title{Cut_branche}
\usage{
cut_branche(gg = "dat", cc = "dit")
}
\arguments{
\item{gg}{the pattern you want to recognize in the name of the files you want to work on.}
}
\description{
Cut_branche
}
