% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R, R/cut_one_side.R
\name{cut_one_side}
\alias{cut_one_side}
\alias{cut_one_side}
\title{cut_one_side}
\usage{
cut_one_side(ii = "dat", gg = "dir")

cut_one_side(ii = "dat", gg = "dir")
}
\arguments{
\item{ii}{pattern in files you want to cut}

\item{dir}{side where you want to delete part of the profile}

\item{ii}{pattern in files you want to cut}

\item{dir}{side where you want to delete part of the profile : "l" left , r "right"}
}
\value{
a profil without the cut part

a profil without the cut part
}
\description{
cut_one_side

cut_one_side
}
\details{
You have to click close to the limit you want to be the end of the profile

You have to click close to the limit you want to be the end of the profile
}
