% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_portion.R
\name{cut_portion}
\alias{cut_portion}
\title{Cut_portion}
\usage{
cut_portion(ii = "dat")
}
\arguments{
\item{ii}{pattern in files you want to cut}
}
\value{
a profil without the cut part
}
\description{
Cut_portion
}
\details{
You have to click two times on the profil to select the area to cut
}
