% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_reverse_CW.R
\name{cut_reverse_CW}
\alias{cut_reverse_CW}
\title{cut_reverse_CW}
\usage{
cut_reverse_CW(ii = "dat")
}
\arguments{
\item{ii}{pattern in files you want to cut}
}
\value{
two separate profil
}
\description{
cut_reverse_CW
}
\details{
You have to click where you want to slip the profil
}
