% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_reverse_OW.R
\name{cut_reverse_OW}
\alias{cut_reverse_OW}
\title{cut_reverse_OW}
\usage{
cut_reverse_OW(ii = "dat")
}
\arguments{
\item{ii}{pattern in files you want to cut}
}
\value{
two separate profil
}
\description{
cut_reverse_OW
}
\details{
You have to click where you want to slip the profil
}
