% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{Import}
\alias{Import}
\title{Import}
\usage{
Import(zz = "dat", pathway = ".")
}
\arguments{
\item{zz}{the pattern you want to recognize in the name of the files you want to import.}

\item{pathway}{path to the files}
}
\description{
Import
}
\details{
to import profiles in R; one by one pressing enter at each new profile.
}
\examples{
Import("pxb"), will import all the profile with pxb in their name.
}
