% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.tot.R
\name{Import.tot}
\alias{Import.tot}
\title{Import.tot}
\usage{
Import.tot(zz = "dat", pathway = ".")
}
\arguments{
\item{zz}{the pattern you want to recognize in the name of the files you want to import.}

\item{pathway}{path to the files}
}
\description{
Import.tot
}
\details{
to import profiles in R.
}
\examples{
Import.tot("pxb"), will import in a raw all the profile with pxb in their name.
}
