% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.bibf.tot.R
\name{k.bibf.tot}
\alias{k.bibf.tot}
\title{K.bibf.tot}
\usage{
k.bibf.tot(tt = "dat")
}
\arguments{
\item{tt}{the pattern you want to recognize in the name of the files you want to to work on.}
}
\description{
K.bibf.tot
}
\details{
position the limite between earlywood and latewood. Create a new object adding a "x" at the beginning of the original name
}
