% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.bibtbf.R
\name{k.bibtbf}
\alias{k.bibtbf}
\title{K.bibtbf}
\usage{
k.bibtbf(tt = "dat")
}
\arguments{
\item{tt}{the pattern you want to recognize in the name of the files you want to work on.}
}
\description{
K.bibtbf
}
\details{
position the limite between earlywood, transition-wood and latewood. Create a new object adding a "t" at the beginning of the original name.
}
