% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supprime.R
\name{supprime}
\alias{supprime}
\title{Supprime}
\usage{
supprime(oo = "dat")
}
\arguments{
\item{oo}{the pattern you want to recognize in the name of the files you want to work on.}
}
\description{
Supprime
}
